ActionUtils: Add screen on action
Add an action to force the screen to turn on. This is a quick
and effective way to turn on the screen when we need it to. Intead
of hijacking WindowManagerPolicy in PWM, we instead call on PM to
do this.

- Make sure 'screen off' action doesn't work while screen is off.
This action should not be called if the screen is already off.

Change-Id: Icacb721ebc6493e4bd0db9d02ac10d66f22eb143
Signed-off-by: alanndz <alanndz7@gmail.com>